feat: add shared transcript segment accumulator (#82)
